requirements.txt
Django==2.1
django-cors-headers==3.1.0
djangorestframework==3.10.3
pytz==2019.2
sqlparse==0.3.0
